Explicitly convert 32-bit integers to 16-bit.
GCC 4.7.0 warns that C++11 considers doing this implicitly ill-formed.

+++ src/asfont.h
@@ -41,7 +41,7 @@ private:
 	void writeLine(Surface *surface, const std::vector<std::string> &text, int x, int y, HAlign halign, VAlign valign);
 
 	SDL_Surface *surface;
-	std::vector<unsigned> charpos;
+	std::vector<Uint16> charpos;
 	std::string characters;
 	int lineHeight;
 };
